Add tests for main handler
