+++ precode_test.go
@@ -1,0 +1,37 @@
+package main
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
+	totalCount := 4
+	req := httptest.NewRequest("GET", "/cafe?count=8&city=moscow", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	require.Equal(t, responseRecorder.Code, http.StatusOK)
+	require.NotEmpty(t, responseRecorder.Body)
+
+	assert.Len(t, strings.Split(responseRecorder.Body.String(), ","), totalCount)
+
+}
+
+func TestMainHandlerWhenCityIsWrong(t *testing.T) {
+	req := httptest.NewRequest("GET", "/cafe?count=3&city=tula", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	require.Equal(t, responseRecorder.Code, http.StatusBadRequest)
+	assert.Equal(t, responseRecorder.Body.String(), "wrong city value")
+}
